gdb: Add qMemTagAddrCheck packet
This commit adds a new packet, qMemTagAddrCheck, allowing GDB remote
targets to use it to query the stub if a given address is tagged.

It also adds a new GDB remote feature, 'memory-tagging-check-addr',
which must be advertised by the GDB stub to inform it can reply to
memory tagging address checks via the new qMemTagCheckAddr packet.

Currently, this address check is done via a read query, where the
contents of /proc/<PID>/smaps is read and the flags in there are
inspected for memory tagging-related flags that indicate the address is
in a memory tagged region.

This is not ideal, for example, for QEMU stub and other cases, like in
baremetal, where there is no notion of any OS file like smaps. Hence,
qMemTagCheckAddr packet allows checking addresses in an OS-agnostic way.

For supporting the new packet, a new target hook is introduced,
target_is_address_tagged, which is used instead of the
gdbarch_tagged_address_p gdbarch hook in the upper layer (printcmd.c).

This new target hook is then specialized per target, for remote.c,
aarch64-linux-nat.c, and corelow.c (the current targets that support
memory tagging).

The target hook in remote.c uses the qMemTagCheckAddr packet to check
an address if the stub advertises the 'memory-tagging-check-add+'
feature in qSupport query, otherwise it falls back to using the current
mechanism, which reads the contents of /proc/<PID>/smaps via vFile
requests.

In the aarch64-linux-nat.c and corelow.c the target hooks use the
gdbarch_tagged_address_p gdbarch hook, so there is no change regarding
how an address is checked in these targets. Just the
gdbarch_tagged_address_p signature is changed for convenience, since
target_check_memtag_addr takes the address to be checked as a CORE_ADDR
type.

Signed-off-by: Gustavo Romero <gustavo.romero@linaro.org>