gdb: Add new remote packet to check if address is tagged
This commit adds a new packet qMemTagAddrCheck allowing GDB remote
targets to use it to query gdbservers if a given address is tagged.

It also adds a new GDB remote feature, 'memory-tagging-check-add+',
which must be advertised by the GDB servers to inform GDB they can reply
to address checks via the new qMemTagAddrCheck remote packet.

Currently, this address check is done via a read query, where the
contents of /proc/<PID>/smaps is read and the flags in there are
inspected for MTE-related flags that indicate the address is in a tagged
memory region.

This is not ideal, for example, on QEMU gdbstub and in other cases,
like in baremetal debugging, where there is no notion of any OS file
like smaps. Hence, qMemTagAddrCheck packet allows check addresses in
an OS-agnostic way.

For supporting the new packet, a new target hook is introduced,
check_memtag_addr, which is used instead of the gdbarch_tagged_address_p
gdbarch hook in the upper layers (printcmd.c).

The new target hook is then specialized per target, for remote.c,
aarch64-linux-nat.c, and corelow.c targets (the current targets that
are MTE-aware).

The target hook in remote.c uses the qMemTagAddrCheck packet to check
an address if the server advertised the 'memory-tagging-check-add+'
feature, otherwise it falls back to using the current mechanism, i.e. it
reads the /proc/<PID>/smaps contents.

In the aarch64-linux-nat.c and corelow.c the target hook uses the
gdbarch_tagged_address_p gdbarch hook, so there is no change regarding
how an address is checked in these targets. Just the
gdbarch_tagged_address_p signature is changed for convenience, since
target_check_memtag_addr takes the address to be checked as a CORE_ADDR
type.

Signed-off-by: Gustavo Romero <gustavo.romero@linaro.org>