gdb: Add qMemTagAddrCheck packet
This commit adds a new packet, qMemTagAddrCheck, allowing GDB remote
targets to use it to query the stub if a given address is tagged.

It also adds a new GDB remote feature, 'memory-tagging-check-addr',
which must be advertised by the GDB stub to inform it can reply to
memory tagging address checks via the new qMemTagCheckAddr packet.

Currently, the memory tagging address check is done via a read query,
where the contents of /proc/<PID>/smaps is read and the flags are
inspected for memory tagging-related flags that indicate the address is
in a memory tagged region.

This is not ideal, for example, for QEMU stub and other cases, such as
on bare-metal, where there is no notion of an OS file like 'smaps.'
Hence, the qMemTagCheckAddr packet allows checking addresses in an
OS-agnostic way.

The is_address_tagged target hook in remote.c uses the qMemTagCheckAddr
packet to check an address if the stub advertises the
'memory-tagging-check-add+' feature, otherwise it falls back to using
the current mechanism, which reads the contents of /proc/<PID>/smaps via
vFile requests.

Signed-off-by: Gustavo Romero <gustavo.romero@linaro.org>